real time testing
I created a Python script to make model inferences on frames collected from the camera. the result of this real-time is fairly bad. The model was not able to detect a face properly. This might be because therer is a difference in the environment of collected  data and test environment.

Again, this is just a preliminary test to check if the pipeline works. Will need to do further training for any meaningful model