add more failed swingup training
for run 15 to 22, the probability of the pendulum up initial state was 40% and the agent only learned to exploit the rewards earned in the pendulum-up episodes.
For run 23 to 32, I reduced this probability to 10%, expecting that the agent tries to explore the swingup motion. The same set of hyperparameters was used to train with 5 random seeds, each seed twice. The training results were horrendous 👎