train with lower learning rates and update rates
lowered the learning rate of critics and actor, and the Polyak coeff, by a factor of 10 for more stable network update. I noticed that the agent was able to achieve approx 500 units of rewards with these lower learning rate and decide to decrease the pass_score to evaluate the models of lower quality.

Policy 66 and 69 were successful at the swingup task, but was not able to bring the cart back to the zero position. This suggests that the pole angle reward has a higher impact on the performance (reward) of the agent than the cart position. I need to experiment with more reward engineering/shaping to improve the training results