add one set of 5x2 experiment with an easier early stopping criterion (2 consecutive passes > 10,000 rewards)
Despite the higher pass score standard, the training policy only need to surpass this value for two consecutive training steps. This resulted in a higher training success rate compared to the last set of experiment.

However, I noticed that the standard deviation of the policy evaluation is a useful insight into whether a policy is actually good. Since there is a low chance that the pendulum start at the pendulum up state, policies that can only do pole balancing but not swing up can still get high rewards from these occurrences. However, this will lead to a high standard deviation in the reward across these evaluation rollouts. The stdev can therefore be used for signalling early stop instead of the mean evaluation reward. When I ran this training set and printed out the evaluation results, the good policies obtained average rewards of ~10k with a standard deviation of <10% of this mean. In the next training iteration, I will change the early stopping criterion from the reward mean to the ratio of reward stdev/mean