add new trial of half cheetah with hyperparams suggested in RLzoo
retrained halfcheetah with TD3 with modified hyperparams
- 1e-3 learning rates (I changed this to make learning faster, does come with some instability though)
- 2e5 buffer size,
- 1e4 learning start,
- architecture of 2 layers 256 nodes each

the result was slightly better compared to the previous run, achieving ~6,000 reward at the end of 500,000 steps, taking about ~30 minutes on the lab hardware. I need to increase the training budget to 1e6 steps to observe the result. I also need to test the MA filter on this training